Code almost complete

+++ Week6/sql/institution_xref_active_in_current_year.2.create table.sql
@@ -2,6 +2,6 @@
             CREATE TABLE institution_xref_active_in_current_year (
                 Codevalue TINYINT,
                 valueLabel VARCHAR(255),
-                valueOrder SMALLINT UNSIGNED,
+                valueOrder MEDIUMINT UNSIGNED,
                 PRIMARY KEY (Codevalue)
             );
