+++ Week6/sql/institution_xref_enrolled_fulltime_firsttime_students.2.create table.sql
@@ -2,6 +2,6 @@
             CREATE TABLE institution_xref_enrolled_fulltime_firsttime_students (
                 Codevalue TINYINT,
                 valueLabel VARCHAR(255),
-                valueOrder SMALLINT UNSIGNED,
+                valueOrder MEDIUMINT UNSIGNED,
                 PRIMARY KEY (Codevalue)
             );
