+++ Week6/sql/institution_xref_multi_inst_id.2.create table.sql
@@ -2,6 +2,6 @@
             CREATE TABLE institution_xref_multi_inst_id (
                 Codevalue VARCHAR(6),
                 valueLabel VARCHAR(255),
-                valueOrder SMALLINT UNSIGNED,
+                valueOrder MEDIUMINT UNSIGNED,
                 PRIMARY KEY (Codevalue)
             );
