+++ Week6/sql/institution_xref_multi_inst_org.2.create table.sql
@@ -2,6 +2,6 @@
             CREATE TABLE institution_xref_multi_inst_org (
                 Codevalue TINYINT,
                 valueLabel VARCHAR(255),
-                valueOrder SMALLINT UNSIGNED,
+                valueOrder MEDIUMINT UNSIGNED,
                 PRIMARY KEY (Codevalue)
             );
