+++ Week6/sql/institution_xref_state.2.create table.sql
@@ -2,6 +2,6 @@
             CREATE TABLE institution_xref_state (
                 Codevalue VARCHAR(2),
                 valueLabel VARCHAR(255),
-                valueOrder SMALLINT UNSIGNED,
+                valueOrder MEDIUMINT UNSIGNED,
                 PRIMARY KEY (Codevalue)
             );
